Make config objects json serializable
