+++ deepspeed/runtime/config_utils.py
@@ -5,8 +5,19 @@ Licensed under the MIT license.
 """
 Collection of DeepSpeed configuration utilities
 """
-
+import json
 from collections import Counter
+
+
+class DeepSpeedConfigObject(object):
+    """
+    For json serialization
+    """
+    def repr(self):
+        return self.__dict__
+
+    def __repr__(self):
+        return json.dumps(self.__dict__, sort_keys=True, indent=4)
 
 
 def get_scalar_param(param_dict, param_name, param_default_value):
