+++ deepspeed/runtime/zero/config.py
@@ -3,13 +3,12 @@ Copyright (c) Microsoft Corporation
 Licensed under the MIT license.
 """
 
-from deepspeed.runtime.config_utils import get_scalar_param
+from deepspeed.runtime.config_utils import get_scalar_param, DeepSpeedConfigObject
 from deepspeed.utils import logger
 from deepspeed.runtime.zero.constants import *
-import json
 
 
-class DeepSpeedZeroConfig(object):
+class DeepSpeedZeroConfig(DeepSpeedConfigObject):
     def __init__(self, param_dict):
         super(DeepSpeedZeroConfig, self).__init__()
 
@@ -65,16 +64,6 @@ class DeepSpeedZeroConfig(object):
             'DeepSpeedConfig: this format of ZeRO optimization setup is deprecated. Please use the following format: {}'
             .format(ZERO_FORMAT))
         return zero_config_dict
-
-    """
-    For json serialization
-    """
-
-    def repr(self):
-        return self.__dict__
-
-    def __repr__(self):
-        return json.dumps(self.__dict__, sort_keys=True, indent=4)
 
     def _initialize(self, zero_config_dict):
         self.stage = get_scalar_param(zero_config_dict,
